\name{canvasSave-methods}
\docType{methods}
\alias{canvasSave-methods}
\alias{canvasFetch-methods}
\alias{canvasSave,rangeMap-method}
\title{Methods to save and fetch the canvas of a \code{rangMapper} project }
\description{
	The canvas is a  regular grid of a given resolution used to perform grid interpolation.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "rangeMap")}}{
%%  ~~describe this method here~~
}

}}
\keyword{methods}

