% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapFetch-methods.R
\name{as.rmap.frame}
\alias{as.rmap.frame}
\alias{as.rmap.frame.data.table}
\title{Convert data.table to rmap.frame}
\usage{
as.rmap.frame(x, ...)

\method{as.rmap.frame}{data.table}(x, p4s, gridSize, bbox)
}
\arguments{
\item{x}{a data.table}

\item{\dots}{extra arguments}

\item{p4s}{proj4string}

\item{gridSize}{grid size}

\item{bbox}{global bounding box, a list  with x and y}
}
\value{
an rmap.frame object which inherits
                from \code{\link[data.table]{data.table}}
}
\description{
Convert data.table to rmap.frame
}
