\name{canvasSave-methods}
\docType{methods}
\alias{canvasSave-methods}
\alias{canvasFetch-methods}
\alias{canvasSave,rangeMap-method}
\title{Methods to save and fetch the canvas of a \code{rangMapper} project }
\description{
	The canvas is a  regular grid of a given resolution. Each range map is overlayed onto the canvas and the results saved to project.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "rangeMap")}}{
The method fails if there is the gridsize is missing or the canvas was allready constructed.
}

}}
\keyword{methods}

