% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dem}
\alias{dem}
\title{dem}
\format{
A RasterLayer with 159 rows and 212 columns.
}
\usage{
data(dem)
}
\description{
Digital elevation model of the Americas based on ETOPO1.
}
\examples{

require(rangeMapper)
data(dem)
raster::plot(dem)

}
\references{
Amante, C. and B. W. Eakins, ETOPO1 1 Arc-Minute Global Relief Model: Procedures, Data Sources and Analysis. 
NOAA Technical Memorandum NESDIS NGDC-24, 19 pp, March 2009. 
Go to this web site: http://www.ngdc.noaa.gov/mgg/global/global.html.
}
\keyword{datasets}
