\name{gui.internals}
\alias{gui.internals}
\alias{gui.make.env}
\alias{gui.get.from.env}
\alias{gui.put.to.env}
\alias{gui.exists.in.env}
\alias{gui.msg}
\alias{gui.img}
\alias{gui.tkEntryBox}
\alias{gui.tkdbBrowse.active.proj}
\alias{gui.help}
\alias{gui.dbopen}
\alias{gui.selectShpFiles}
\alias{gui.global.bbox.save}
\alias{gui.gridSize.save}
\alias{gui.canvas.save}
\alias{gui.processRanges}
\alias{gui.bio.save}
\alias{gui.chooseFunction}
\alias{gui.tkColorPalette}
\alias{gui.rangeMap.save}
\alias{gui.rangeMap.plot}
\alias{gui.yn}
\alias{gui.rangeMap.rm}
\alias{gui.tooltips}
\alias{gui.chooseSubset}
\alias{gui.chooseVariable}
\alias{gui.close}
\alias{gui.tkComboEntryBox}
\alias{gui.show.metadata}

\title{Undocumented gui functions}
\description{These functions are undocumented; All functions starting with 'gui.' are used internally by \code{\link{rangeMapper}}}
\note{For details check the functions without the \sQuote{gui.} prefix}
\author{Mihai Valcu \email{valcu@orn.mpg.de}}