\name{internals}
\alias{internals}
\alias{.sqlQuery}
\alias{.is.empty}
\alias{.extract.p4s}
\alias{sp.metadata}
\alias{richness}
\title{Undocumented functions}
\description{These functions are undocumented; most are internal and not intended for direct use.}
\author{Mihai Valcu \email{valcu@orn.mpg.de}}



