% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{palette_rangemap}
\alias{palette_rangemap}
\title{A few color palettes}
\usage{
palette_rangemap(set = "set1")
}
\arguments{
\item{set}{set type (currently one set only)}
}
\description{
A few color palettes
}

