% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapPlot-methods.R
\docType{methods}
\name{plot,rmap.frame,missing-method}
\alias{plot,rmap.frame,missing-method}
\title{Plot a rmap.frame}
\usage{
\S4method{plot}{rmap.frame,missing}(x)
}
\arguments{
\item{x}{a a rmap.frame object.}
}
\value{
a ggplot object.
}
\description{
Plot a rmap.frame
}

