% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangeMap.R
\docType{class}
\name{rangeMap-class}
\alias{rangeMap}
\alias{rangeMap-class}
\alias{summary,rangeMap-method,}
\title{Class "rangeMap" to formally describe a rangeMapper sqlite file.}
\usage{
rangeMap(path)
}
\description{
This is the basic class of the package and the only one documented and exported.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("rangeMap", \dots{})}.
}
\author{
Mihai Valcu, \email{valcu@orn.mpg.de}
}
\seealso{
\code{\link{rangeMapper}}, \code{\link{processRanges}}
}
\keyword{classes}

