% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapFetch-methods.R
\name{rangeMap.fetch}
\alias{rangeMap.fetch}
\title{rangeMap.fetch
rangeMap.fetch}
\usage{
rangeMap.fetch(con, maps, spatial = TRUE)
}
\arguments{
\item{spatial}{}
}
\description{
rangeMap.fetch
rangeMap.fetch
}

