% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectINI-methods.R
\name{rangeMap.start}
\alias{rangeMap.open}
\alias{rangeMap.start}
\title{Initiate/open a new rangeMapper project}
\usage{
rangeMap.start(...)

rangeMap.open(path, verbose = TRUE)
}
\arguments{
\item{path}{Character vector; a path to a valid rangeMapper project.}

\item{verbose}{Character vector; if \code{TRUE} the project's summary is printed.}

\item{file}{Project's file name.}

\item{dir}{Project directory.}

\item{overwrite}{Logical vector, default to FALSE (the file is kept but all tables are dropped).}
}
\value{
rangeMap.start() and rangeMap.open() returns an sqlite connection.
}
\description{
Initiate/open a new rangeMapper project
}
\examples{

td = setwd(tempdir())

dbcon = rangeMap.start(file = "test.sqlite", overwrite = TRUE, dir = tempdir() )
summary(dbcon)

summary(rangeMap("test.sqlite"))

dbcon = rangeMap.open(path = "test.sqlite")
summary(dbcon)
setwd(td)


}
\seealso{
\code{\link{rangeMap.save}}
}

