
\name{rangeMapBbox}
\alias{rangeMapBbox}
\alias{rangeMapBbox-methods}
\alias{rangeMapBbox-method}
\alias{rangeMapBbox,rangeFiles-method}

\alias{rangeMapBboxFetch-methods}
\alias{rangeMapBboxFetch}
\alias{rangeMapBboxFetch,rangeMapBboxFetch-method}
\alias{rangeMapBboxFetch,rangeMap-method}

\alias{rangeMapBboxSave-methods}
\alias{rangeMapBboxSave}
\alias{rangeMapBboxSave,rangeMapBboxSave-method}
\alias{rangeMapBboxSave,rangeMap,character,CRS-method}
\alias{rangeMapBboxSave,rangeMap,character,missing-method}
\alias{rangeMapBboxSave,rangeMap,missing,CRS-method}
\alias{rangeMapBboxSave,rangeMap,missing,missing-method}
\alias{rangeMapBboxSave,rangeMap,Spatial,missing-method}




\docType{methods}


\title{ Methods for Function rangeMapBbox in Package `rangeMapper' }
\description{
Methods for function \code{rangeMapBbox} in Package `rangeMapper' 
}
\section{Methods}{
\describe{

\item{\code{signature(object = "rangeFiles")}}{

}

}}

\author{
Mihai Valcu \email{valcu@orn.mpg.de}
}


\seealso{
	\code{\link{global.bbox}}
}

\keyword{methods}














