\name{rangeMapFetch}
\alias{rangeMapFetch}
\alias{rangeMapFetch-methods}
\alias{rangeMapFetch-class}
\alias{rangeMapFetch,rangeMapFetch-method}

\Rdversion{1.1}
\docType{class}



\title{Class "rangeMapFetch"}
\description{
The methods of this class allows fetching maps from a \code{rangeMapper} project.
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("rangeMapFetch", ...)}.
 
}
\section{Slots}{
  \describe{
	\item{\code{tableName}:}{Object of class \code{"character"} MAP name (without the prefix) to fetch.}
  }
}
\section{Extends}{
Class \code{"\linkS4class{rangeMap}"}, directly.
}
\section{Methods}{
  \describe{
    \item{rangeMapFetch}{\code{signature(object = "rangeMapFetch")}: ... }
	 }
}

\author{ 
	Mihai Valcu \email{valcu@orn.mpg.de} }



\seealso{
\code{\link{rangeMap.fetch}}
\code{\link{SpatialPixelsRangeMap-class}}


}


\keyword{classes}
