% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI.R, R/rangeMapper-package.R
\docType{package}
\name{rangeMapper}
\alias{rangeMapper}
\alias{rangeMapper-package}
\title{rangeMapper GUI}
\usage{
rangeMapper()
}
\description{
rangeMapper GUI

\pkg{rangeMapper} is a front end platform for the study of macroecology of
life history traits at both inter-specific and assemblage levels.
}
\details{
The package uses species range (extent-of-occurrence) vector polygons and
life history traits data-sets to build up maps (e.g. species richness, mean
body mass, \dots{}).
}
\section{Getting Started}{
 For a quick start open the graphical user
interface (\bold{gui}) by typing \code{rangeMapper()}.
}
\examples{
\dontrun{
rangeMapper()
}


\dontrun{
rangeMapper()
}

}
\author{
Mihai Valcu \email{valcu@orn.mpg.de},
\url{http://orn.mpg.de/valcu.html} \cr James Dale
\email{J.Dale@massey.ac.nz}, \url{http://quelea.net}
}
\references{
Valcu, M., Dale, J. and Kempenaers, B. (2012) rangeMapper: A
platform for the study of macroecology of life history traits. 21(9). (DOI:
10.1111/j.1466-8238.2011.00739.x)
}
\seealso{
\code{\link[rangeMapper]{rangeMap.save}}.
}
\keyword{gui}
\keyword{macroecology}
\keyword{package}
\keyword{spatial}
\keyword{tcltk}

