% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapProcess-methods.R
\docType{methods}
\name{rangeMapProcess}
\alias{rangeMapProcess}
\alias{rangeMapProcess,rangeMapProcess,SpatialPolygonsDataFrame,missing,character,list,logical-method}
\alias{rangeMapProcess,rangeMapProcess,SpatialPolygonsDataFrame,missing,character,missing,logical-method}
\alias{rangeMapProcess,rangeMapProcess,missing,character,missing,list,missing-method}
\alias{rangeMapProcess,rangeMapProcess,missing,character,missing,missing,missing-method}
\title{rangeMapProcess}
\usage{
rangeMapProcess(object, spdf, dir, ID, metadata, parallel)


  \S4method{rangeMapProcess}{rangeMapProcess,missing,character,missing,missing,missing}(object,
  dir)


  \S4method{rangeMapProcess}{rangeMapProcess,missing,character,missing,list,missing}(object,
  dir, metadata)


  \S4method{rangeMapProcess}{rangeMapProcess,SpatialPolygonsDataFrame,missing,character,missing,logical}(object,
  spdf, ID, metadata, parallel)


  \S4method{rangeMapProcess}{rangeMapProcess,SpatialPolygonsDataFrame,missing,character,list,logical}(object,
  spdf, ID, metadata, parallel)
}
\arguments{
\item{object}{a \code{rangeMapProcess} object.}

\item{spdf}{\code{\link[sp]{SpatialPolygonsDataFrame}} object containing all the ranges.}

\item{dir}{ranges file directory where the individual ranges shp files are located. In this case the range ID is the file name.}

\item{ID}{when spdf is set this is a \code{character} vector given the name of the range.}

\item{metadata}{a named list of functions. See \code{\link[rangeMapper]{rangeTraits}} and \code{\link[rangeMapper]{metadataUpdate}}}

\item{parallel}{logical, default to false.}
}
\description{
rangeMapProcess

Method 1: each range file is a separate shp file. No metadata

Method 2: Each range file is a separate shp file. Metadata are computed

Method 3: One shp file containing all ranges, ID is required. No metadata

Method 4: One shp file containing all ranges, ID is required.  Metadata are computed
}
\seealso{
\code{\link[rangeMapper]{processRanges}}
}

