% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapShow-methods.R
\name{rangeMapProjInfo}
\alias{rangeMapProjInfo}
\title{rangeMap file info}
\usage{
rangeMapProjInfo(con)
}
\arguments{
\item{con}{a connection to a rangeMapper project.}
}
\value{
data.table
}
\description{
rangeMap file info
}

