\name{rangeMapRemove}
\alias{rangeMapRemove}
\alias{rangeMapRemove-methods}
\alias{rangeMapRemove-class}
\alias{rangeMapRemove,rangeMapRemove-method}

\Rdversion{1.1}
\docType{class}

\title{Class "rangeMapRemove"}
\description{
The method defined for this class allows removing \code{rangeMapper} objects.
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("rangeMapRemove", ...)}.
}

\section{Slots}{
  \describe{
    \item{\code{tableName}:}{Object of class \code{"character"}  }
	\item{\code{tablePrefix}:}{Object of class \code{"character"} `MAP' or `BIO' }

  }
}
\section{Extends}{
Class \code{"\linkS4class{rangeMap}"}, directly.
}

\section{Methods}{
  \describe{
    \item{rangeMapRemove}{\code{signature(object = "rangeMapRemove")}: ... }
	 }
}

\author{Mihai Valcu \email{valcu@orn.mpg.de}}


\seealso{\code{\link{rm.rangeMapper}} }

\keyword{classes}
