% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-spatial.R
\name{rangeOverlay}
\alias{rangeOverlay}
\title{rangeOverlay}
\usage{
rangeOverlay(spp, canvas, name)
}
\arguments{
\item{spp}{a SpatialPolygons* object.}

\item{canvas}{a SpatialPointsDataFrame.}

\item{name}{a character vector (a bioid)}
}
\value{
a data.frame with two columns: id (canvas id) and bioid.
}
\description{
rangeOverlay
}
