% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapRemove-methods.R
\name{rm.rangeMapper}
\alias{rm.rangeMapper}
\title{Remove tables from a give project}
\usage{
rm.rangeMapper(con, ...)
}
\arguments{
\item{con}{A valid sqlite connection.}

\item{\dots}{Arguments passed to the corresponding methods specifically
\sQuote{tablePrefix} or \sQuote{tableName}}
}
\description{
Remove tables given prefix attribute or by name
}
\note{
The default \sQuote{rm.rangeMapper(con)} will remove all \sQuote{MAP}
              and \sQuote{BIO} tables.
}
