% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{st_thin}
\alias{st_thin}
\title{Thinning of polygonal grids}
\usage{
st_thin(x, order)
}
\arguments{
\item{x}{an sf data.frame}

\item{lag}{lag order.}
}
\value{
what type of object does it return.
}
\description{
Nearest neighbours spatial thinning of polygonal grids
}
\examples{
require(rangeMapper)
require(data.table)
con = rmap_connect()
data(wrens)
rmap_add_ranges(con, x = wrens, ID = 'sci_name')
rmap_prepare(con, 'hex', cellsize=250)
rmap_save_map(con) 
x = rmap_to_sf(con)[, 'cell_id']

x = x[x$cell_id != 356, ]
ggplot(x) + geom_sf() + geom_sf_label( aes(label = cell_id))

plot( st_thin(x,5) \%>\% st_geometry )

}
\references{
Based on SO answer: https://stackoverflow.com/questions/65907022/
}
