% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapShow-methods.R
\docType{methods}
\name{tables,SQLiteConnection-method}
\alias{tables,SQLiteConnection-method}
\title{Tables and column names  of an sqlite db}
\usage{
\S4method{tables}{SQLiteConnection}(object)
}
\arguments{
\item{object}{an sqlite connection object.}
}
\value{
data.frame
}
\description{
Tables and column names  of an sqlite db
}
