% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{theme_rangemap}
\alias{theme_rangemap}
\title{ggplot theme}
\usage{
theme_rangemap(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{base_size}

\item{base_family}{base_family}
}
\description{
A ggplot theme based on \code{\link[ggplot2]{theme_bw}}
}

