% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI.R
\name{View_rmap}
\alias{View_rmap}
\title{rangeMapper browser}
\usage{
View_rmap(path)
}
\arguments{
\item{path}{path to a rangeMapper project. If missing a demo project is created on the fly.}
}
\description{
rangeMapper browser
}
\examples{
\dontrun{
View_rmap()
}

}
