
\name{brewer.pal.get}
\alias{brewer.pal.get}


\title{
  brewer.pal color palettes.
}
\description{
 A simple wrapper around \code{\link{brewer.pal.info}}
}
\usage{
brewer.pal.get(palette)
}
\arguments{
  \item{palette}{The name of a color palette. When NULL all the qualitative color palettes are returned.}

}



\author{Mihai Valcu \email{valcu@orn.mpg.de}}

\seealso{
  \code{\link{tkColorPalette}}
  \code{\link{brewer.pal}}
  
}


\keyword{color}


