% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapPlot-methods.R
\docType{methods}
\name{plot,SpatialPixelsRangeMap,missing-method}
\alias{plot,SpatialPixelsRangeMap,missing-method}
\title{Plot a SpatialPixelsRangeMap}
\usage{
\S4method{plot}{SpatialPixelsRangeMap,missing}(x,
  colorpalette = brewer.pal.get("Spectral")[11:1], ncols = 20,
  scales = FALSE, style = "equal", ...)
}
\arguments{
\item{x}{A SpatialPixelsRangeMap}

\item{colorpalette}{A color palette.}

\item{ncols}{Number of color classes required, default to 20; argument to be passed to
\code{\link[classInt]{classIntervals}}}

\item{scales}{If \sQuote{FALSE}, default, axes scale are not drawn.}

\item{style}{Class interval style; see \code{\link[classInt]{classIntervals}} for more details}

\item{\dots}{Any argument that can be passed to see \code{\link[sp]{spplot}}}
}
\description{
This is a wrapper around \code{spplot}
}
\examples{
require(rangeMapper)
require(rgdal)
dbcon = rangeMap.start(file = "test.sqlite", overwrite = TRUE, dir = tempdir() )
f = system.file(package = "rangeMapper", "extdata", "wrens", "vector_combined")
global.bbox.save(con = dbcon, bbox = f,
p4s = CRS("+proj=cea +lon_0=0 +lat_ts=30 +x_0=0 +y_0=0 +ellps=WGS84 +units=km +no_defs") )
gridSize.save(dbcon, gridSize = 500)  # cell size 2 deg
canvas.save(dbcon)
processRanges(spdf = readOGR(f, "wrens", verbose = FALSE), con =  dbcon, ID = "sci_name")
rangeMap.save(dbcon) # species richness


# PLOTS
all = rangeMap.fetch(dbcon)
SR = rangeMap.fetch(dbcon, 'species_richness')

plot(all)
plot(SR, style = "fisher", sub = "Wrens species richness")

pal =  RColorBrewer::brewer.pal(11, 'RdYlGn')[11:1]

plot(SR, style = "fisher", colorpalette = pal)

}

