% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapExport-methods.R
\name{rangeMap.export}
\alias{rangeMap.export}
\title{Export \sQuote{MAP} tables}
\usage{
rangeMap.export(con, dirName = dirName, ...)
}
\arguments{
\item{con}{An sqlite connection pointing to a valid \code{rangeMapper}
project.}

\item{dirName}{The directory name where the \sQuote{MAP}s will be exported.
If missing, the \sQuote{MAP}s will be exported in project's directory}

\item{\dots}{Further arguments to pass to \code{\link{writeGDAL}}}
}
\description{
Export \sQuote{MAP} tables as single-band geotiff files
}

