% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapFetch-methods.R
\name{rangeMap.fetch}
\alias{rangeMap.fetch}
\title{rangeMap.fetch}
\usage{
rangeMap.fetch(con, maps, spatial = TRUE)
}
\arguments{
\item{con}{a connection to a valid rangeMapper project.}

\item{maps}{map(s) name as character vector. If missing then all the maps are returned.}

\item{spatial}{If TRUE (default) a \code{SpatialPixelsRangeMap} is returned, else a \code{rmap.frame}.}
}
\value{
an object of SpatialPixelsRangeMap or \code{data.table}  containing the spatial coordinates
               and \code{proj4} string as an attribute if \code{spatial = FALSE}.
}
\description{
rangeMap.fetch
}
