% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangeMapper-package.R
\docType{package}
\name{rangeMapper}
\alias{rangeMapper}
\alias{rangeMapper-package}
\title{rangeMapper: A platform for the study of macroecology of life history
traits.}
\description{
\pkg{rangeMapper} is a front end platform for the study of macroecology of
life history traits at both inter-specific and assemblage levels.
}
\references{
Valcu, M., Dale, J. and Kempenaers, B. (2012) rangeMapper: A
     platform for the study of macroecology of life history traits. 21(9). (DOI:
     10.1111/j.1466-8238.2011.00739.x)
}
