% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{rangeMapper}
\alias{rangeMapper}
\alias{rangeMapper-package}
\title{A Platform for the Study of Macro-Ecology of Life History Traits.}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools for generation of (life-history) traits and diversity maps on hexagonal or square grids. Valcu et al.(2012) \doi{10.1111/j.1466-8238.2011.00739.x}.
}
\section{Package options}{


rangeMapper global options can be changed with options(name = value):

\itemize{
  \item `rmap.verbose`: defaults to TRUE. Setting this option to FALSE will shut up all custom messages.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mpio-be/rangeMapper}
  \item Report bugs at \url{https://github.com/mpio-be/rangeMapper/issues}
}

}
\author{
\strong{Maintainer}: Mihai Valcu \email{valcu@orn.mpg.de} (\href{https://orcid.org/0000-0002-6907-7802}{ORCID})

Authors:
\itemize{
  \item James Dale \email{J.Dale@massey.ac.nz} (\href{https://orcid.org/0000-0001-5950-5829}{ORCID})
}

Other contributors:
\itemize{
  \item Joan Maspons [contributor]
}

}
\keyword{internal}
