% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{st_thin}
\alias{st_thin}
\title{Thinning of polygonal grids}
\usage{
st_thin(x, lag)
}
\arguments{
\item{x}{an sf data.frame.}

\item{lag}{lag order.}
}
\value{
a thinned \code{\link[sf:st_as_sf]{sf::st_as_sf()}} object.
}
\description{
Nearest neighbours spatial thinning of polygonal grids
}
\note{
This function is still under development.
}
\examples{
\dontrun{
require(rangeMapper)
con = rmap_connect()
wrens = read_wrens()
rmap_add_ranges(con, x = wrens, ID = 'sci_name')
rmap_prepare(con, 'hex', cellsize=500)
rmap_save_map(con) 
x = rmap_to_sf(con)[, 'cell_id']

plot( st_thin(x,2) )

x = x[ ! x$cell_id  \%in\%  c(282,265) , ]

plot( st_thin(x,3) )

}
}
\references{
Based on SO answer: https://stackoverflow.com/questions/65907022/
}
